Module Testo_diff

A simple diffing algorithm

The current implementation is a copy of Gabriel Jaldon's original OCaml implementation at https://github.com/gjaldon/simple-diff

The heuristic being used is subject to change and may result in slightly different results in future versions.

module type Comparable = sig ... end

module type S = sig ... end

module Make (Item : Comparable) : S with type item = Item.t