Module Testo_diff.Make

Parameters

module Item : Comparable

Signature

type item = Item.t

The type of the item that will be compared.

val pp_item : Ppx_deriving_runtime.Format.formatter -> item -> Ppx_deriving_runtime.unit

val show_item : item -> Ppx_deriving_runtime.string

type diff =

| Deleted of item array

| Added of item array

| Equal of item array

Represents the change or lack of change in a line or character between the old and new version.

val pp_diff : Ppx_deriving_runtime.Format.formatter -> diff -> Ppx_deriving_runtime.unit

val show_diff : diff -> Ppx_deriving_runtime.string

type t = diff list

List of diffs which is the return value of the main function.

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val get_diff : item array -> item array -> t

Returns a list of diffs between two arrays

val recover_input : t -> item array * item array

Recover the original input passed to get_diff.